Remove syntex-based code generation.
Now that Rust 1.15 is released, custom derive via Macros 1.1 is stable,
so the need to split code generation into serde_derive on nightly Rust
and serde_codegen on stable Rust is no longer needed.

Serde has also been updated to the latest 0.9 series release. There were
breaking changes to Serde's API, but nothing that affects consumers of
Rusoto.

Finally, and not really related, test configuration and documentation
suggesting installation of OpenSSL on macOS and Windows has been
removed, as we now use Hyper 0.10, which no longer requires OpenSSL on
these platforms.

Closes #555.
