+++ README.md
@@ -31,9 +31,9 @@ You may be looking for:
 
 ## Requirements
 
-Rust 1.13.0 or later is required.
+Rust 1.15.0 or later is required.
 
-On OS X and Windows, you may need to install the openssl runtime and headers to get the `rust-openssl` dependency to build. Instructions for that can be found [here](https://github.com/sfackler/rust-openssl#building).
+On Linux, OpenSSL is required.
 
 ## Installation
