+++ appveyor.yml
@@ -23,12 +23,6 @@ install:
 
 # Equivalent to Travis' `script` phase
 build_script:
-  # https://github.com/sfackler/rust-openssl/issues/293
-  - set DEP_OPENSSL_INCLUDE=C:\OpenSSL-Win64\include
-  - set OPENSSL_INCLUDE_DIR=C:\OpenSSL-Win64\include
-  - set OPENSSL_LIBS=ssleay32MT:libeay32MT
-  - set OPENSSL_LIB_DIR=C:\OpenSSL-Win64\lib\VC
-  - path C:\OpenSSL-Win64\bin;%PATH%
   # We only build and test a subset due to https://github.com/rusoto/rusoto/issues/537
   - cargo build --features "dynamodb s3 sqs" -vv
 test_script:
