+++ codegen/src/lib.rs
@@ -1,7 +1,5 @@
-#![cfg_attr(feature = "unstable", feature(const_fn, drop_types_in_const))]
 #![cfg_attr(feature = "nightly-testing", feature(plugin))]
 #![cfg_attr(feature = "nightly-testing", plugin(clippy))]
-#![cfg_attr(not(feature = "unstable"), deny(warnings))]
 
 extern crate inflector;
 #[macro_use]
@@ -11,13 +9,10 @@ extern crate serde;
 extern crate serde_json;
 extern crate hyper;
 
-#[cfg(feature = "serde_derive")]
 #[macro_use]
 extern crate serde_derive;
-#[cfg(not(feature = "serde_derive"))]
-extern crate serde_codegen;
 
-use std::fs::{File, rename};
+use std::fs::File;
 use std::io::BufReader;
 use std::path::Path;
 
@@ -47,27 +42,17 @@ impl Service {
 }
 
 pub fn generate(service: Service, output_path: &Path) -> i32 {
-    let botocore_destination_path = output_path.join(format!("{}_botocore.rs", service.name));
-    let serde_destination_path = output_path.join(format!("{}.rs", service.name));
+    let destination_path = output_path.join(format!("{}.rs", service.name));
     let botocore_service_data_path = Path::new(BOTOCORE_DIR)
         .join(format!("{}/{}/service-2.json", service.name, service.protocol_date));
 
-    let needs_serde = botocore_generate(botocore_service_data_path.as_path(),
-                                        botocore_destination_path.as_path());
-
-    // only pass the generated code through serde if we actually need JSON serialization
-    if needs_serde {
-        serde_generate(botocore_destination_path.as_path(),
-                       serde_destination_path.as_path());
-    } else {
-        rename(botocore_destination_path, serde_destination_path).unwrap();
-    }
+    botocore_generate(botocore_service_data_path.as_path(),
+                                        destination_path.as_path());
 
     return 1;
-
 }
 
-fn botocore_generate(input_path: &Path, output_path: &Path) -> bool {
+fn botocore_generate(input_path: &Path, output_path: &Path) {
     let input_file = File::open(input_path).expect(&format!(
         "{:?} not found",
         input_path,
@@ -84,23 +69,4 @@ fn botocore_generate(input_path: &Path, output_path: &Path) -> bool {
         Ok(()) => {},
         _ => panic!("Failed to write file at {:?}", output_path)
     }
-
-    match &service.metadata.protocol[..] {
-        "json" | "rest-json" => true,
-        _ => false,
-    }
-}
-
-#[cfg(not(feature = "serde_derive"))]
-fn serde_generate(source: &Path, destination: &Path) {
-    ::serde_codegen::expand(&source, &destination).unwrap();
-}
-
-#[cfg(feature = "serde_derive")]
-fn serde_generate(source: &Path, destination: &Path) {
-    ::std::fs::copy(source, destination).expect(&format!(
-        "Failed to copy {:?} to {:?}",
-        source,
-        destination,
-    ));
 }
