+++ credential/Cargo.toml
@@ -16,12 +16,11 @@ version = "0.4.0"
 chrono = "0.2.25"
 reqwest = "0.3.0"
 regex = "0.2.1"
-serde_json = "0.8.3"
+serde_json = "0.9.4"
 
 [dependencies.clippy]
 optional = true
 version = "0.0"
 
 [features]
-nightly-testing = ["clippy", "unstable"]
-unstable = []
+nightly-testing = ["clippy"]
